Split up the cache test suite so it's easier to understand and extend
